requirements.txt
langchain-core>=0.1.17
langchain-community>=0.0.19
langchain-google-genai>=0.0.8
python-dotenv>=1.0.0
pydantic>=2.5.3
requests>=2.31.0
python-dateutil>=2.8.2
typing-extensions>=4.9.0
duckduckgo-search>=4.4.0
wikipedia>=1.4.0 